backend/requirements.txt
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.2
asttokens==2.0.5
attrs==21.4.0
autopep8==1.7.0
backcall==0.2.0
beautifulsoup4==4.11.1
bitstring==4.0.1
bleach==5.0.0
boto3==1.24.7
botocore==1.27.7
bz2file==0.98
certifi==2022.6.15
cffi==1.15.0
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.2
cycler==0.11.0
debugpy==1.6.0
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.4
Django==4.0.5
django-cors-headers==3.13.0
django-rest-knox==4.2.0
djangorestframework==3.13.1
drf-yasg==1.21.4
dynaconf==3.1.9
ecdsa==0.18.0
entrypoints==0.4
esptool==4.5
et-xmlfile==1.1.0
executing==0.8.3
fastjsonschema==2.15.3
filelock==3.7.0
Flask==2.1.3
fonttools==4.33.3
graphics.py==5.0.1.post1
hvac==0.11.2
idna==3.3
inflection==0.5.1
ipykernel==6.13.0
ipython==8.3.0
ipython-genutils==0.2.0
itsdangerous==2.1.2
itypes==1.2.0
jedi==0.18.1
Jinja2==3.1.2
jmespath==1.0.0
joblib==1.1.0
jsonschema==4.5.1
jupyter-client==7.3.0
jupyter-core==4.10.0
jupyterlab-pygments==0.2.2
kiwisolver==1.4.2
knox==0.1.14
loguru==0.6.0
MarkupSafe==2.1.1
matplotlib==3.5.2
matplotlib-inline==0.1.3
mistune==0.8.4
nbclient==0.6.2
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio==1.5.5
nltk==3.7
notebook==6.4.11
numpy==1.22.3
opencv-python==4.6.0.66
openpyxl==3.0.10
packaging==21.3
paho-mqtt==1.6.1
pandas==1.4.2
pandocfilters==1.5.0
parso==0.8.3
pickleshare==0.7.5
Pillow==9.1.1
platformdirs==2.5.2
prometheus-client==0.14.1
prompt-toolkit==3.0.29
psutil==5.9.0
pure-eval==0.2.2
pycodestyle==2.9.1
pycparser==2.21
Pygments==2.12.0
pynput==1.7.6
pyparsing==3.0.8
pyrsistent==0.18.1
pyserial==3.5
python-dateutil==2.8.2
PythonTurtle==0.3.2
pytz==2022.1
pywin32==304
pywinpty==2.0.5
pyzmq==22.3.0
reedsolo==1.6.0
regex==2022.6.2
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
s3transfer==0.6.0
scikit-learn==1.0.2
scipy==1.8.0
Send2Trash==1.8.0
shortuuid==1.0.9
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
stack-data==0.2.0
terminado==0.13.3
textblob==0.17.1
threadpoolctl==3.1.0
tinycss2==1.1.1
toml==0.10.2
tornado==6.1
tqdm==4.64.0
traitlets==5.1.1
tzdata==2022.1
uritemplate==4.1.1
urllib3==1.26.11
validators==0.20.0
virtualenv==20.14.1
virtualenvwrapper-win==1.2.7
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.2.1
win32-setctime==1.1.0
xlrd==1.2.0
